ftrace: make direct_call a capability
Make direct_call in struct ftrace_ops a capability. It is written to a
register and used as a destination for a jump instruction.

Update all functions that generate or relay values for direct_call.

Don't convert "direct" in struct ftrace_func_entry. An address is
sufficient for hash calculation and symbol lookup.

+++ include/linux/ftrace.h
@@ -442,7 +442,7 @@ struct ftrace_ops {
 	ftrace_ops_func_t		ops_func;
 	struct ftrace_ops		*managed;
 #ifdef CONFIG_DYNAMIC_FTRACE_WITH_DIRECT_CALLS
-	unsigned long			direct_call;
+	uintptr_t			direct_call;
 #endif
 #endif
 };
@@ -520,11 +520,11 @@ struct ftrace_func_entry {
 
 #ifdef CONFIG_DYNAMIC_FTRACE_WITH_DIRECT_CALLS
 unsigned long ftrace_find_rec_direct(unsigned long ip);
-int register_ftrace_direct(struct ftrace_ops *ops, unsigned long addr);
+int register_ftrace_direct(struct ftrace_ops *ops, uintptr_t addr);
 int unregister_ftrace_direct(struct ftrace_ops *ops, unsigned long addr,
 			     bool free_filters);
-int modify_ftrace_direct(struct ftrace_ops *ops, unsigned long addr);
-int modify_ftrace_direct_nolock(struct ftrace_ops *ops, unsigned long addr);
+int modify_ftrace_direct(struct ftrace_ops *ops, uintptr_t addr);
+int modify_ftrace_direct_nolock(struct ftrace_ops *ops, uintptr_t addr);
 
 void ftrace_stub_direct_tramp(void);
 
@@ -534,7 +534,7 @@ static inline unsigned long ftrace_find_rec_direct(unsigned long ip)
 {
 	return 0;
 }
-static inline int register_ftrace_direct(struct ftrace_ops *ops, unsigned long addr)
+static inline int register_ftrace_direct(struct ftrace_ops *ops, uintptr_t addr)
 {
 	return -ENODEV;
 }
@@ -543,11 +543,11 @@ static inline int unregister_ftrace_direct(struct ftrace_ops *ops, unsigned long
 {
 	return -ENODEV;
 }
-static inline int modify_ftrace_direct(struct ftrace_ops *ops, unsigned long addr)
+static inline int modify_ftrace_direct(struct ftrace_ops *ops, uintptr_t addr)
 {
 	return -ENODEV;
 }
-static inline int modify_ftrace_direct_nolock(struct ftrace_ops *ops, unsigned long addr)
+static inline int modify_ftrace_direct_nolock(struct ftrace_ops *ops, uintptr_t addr)
 {
 	return -ENODEV;
 }
@@ -566,7 +566,7 @@ static inline int modify_ftrace_direct_nolock(struct ftrace_ops *ops, unsigned l
  * instead of going back to the function it just traced.
  */
 static inline void arch_ftrace_set_direct_caller(struct ftrace_regs *fregs,
-						 unsigned long addr) { }
+						 uintptr_t addr) { }
 #endif /* CONFIG_DYNAMIC_FTRACE_WITH_DIRECT_CALLS */
 
 #ifdef CONFIG_STACK_TRACER
